requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
passlib[bcrypt]
PyJWT
cachetools
jinja2
python-multipart
torch
transformers
sentence-transformers
scikit-learn